requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
azure-core==1.2.2
azure-storage-blob==12.2.0
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==6.7
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
distro-info===0.18ubuntu0.18.04.1
et-xmlfile==1.0.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
incremental==16.10.1
isodate==0.6.0
jdcal==1.4.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
MarkupSafe==1.0
matplotlib==3.1.3
msrest==0.6.11
netifaces==0.10.4
numpy==1.18.1
oauthlib==3.1.0
openpyxl==3.0.3
PAM==0.4.2
pandas==1.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyparsing==2.4.6
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
scipy==1.4.1
seaborn==0.10.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==16.0.0
six==1.14.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
WALinuxAgent==2.2.45
xlrd==1.2.0
zope.interface==4.7.1
